SplitInWithFoldableValue
